+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Release Notes
 =============
 
+Forthcoming
+-----------
+* [ports] Auto-register ``PortsMixin`` subclasses for the XML parser; ``init_lookup`` is now optional
+* [ports] Add ``register_ports_class()`` / ``get_ports_registry()`` and ``parse_behaviour_tree_xml(..., auto_register=True)``
+
 2.4.0 (2025-11-13)
 ------------------
 * [code] Replace type checks with instance checks (`#479 <https://github.com/splintered-reality/py_trees/issues/479>`_)